[Feat] container component dir structure
